StringUtil: Replace deprecated Html.fromHtml(String) calls
Html.fromHtml(String) method was deprecated in API level 24. A version check
method is added to StringUtil class to use old method on before API level 24
and new method on API level 24 and higher. Additionally, the method is overloaded
to satisfy overloaded version of fromHtml(String) method, which is fromHtml(String, ImageGetter, TagHandler).

Bug: T145582
Change-Id: I256ce2b3a35e5d8434b7d392eb1c3b102f2e7af7