Ellipsize nav tab labels.
It's an unfortunate reality that in some languages, the labels for our
bottom navigation tabs are too long to fit in the small space allowed by
each tab.

This patch allows the labels to be ellipsized, instead of the default
behavior of overflowing onto multiple lines (even though it's set to
maxLines=1).

Bug: T146579
Change-Id: I3dad2f0a53b31261c3f13c6d37508855558a56cc