Always open pages from Intent in a new foreground tab.
Some users, as well as our own dogfooding, have expressed some confusion
around the behavior of the hardware Back button when reading articles.

The confusion seems to stem from the following:
When a user opens a page from within MainActivity (from the feed, History,
etc), it causes PageActivity to open, and load the requested page into the
*current tab* that may have been previously open. This causes some
unexpected behavior:  if the user presses the hardware Back button,
PageActivity will go backwards in the backstack of pages that had already
existed in the current tab. This is contrary to the intuition that the
Back button should simply close the current page and go back to
MainActivity from where we launched it.

This patch attempts to alleviate the confusion by always loading pages in
a new tab (that is, pages that are loaded from an Intent that causes
PageActivity to start). Here are the benefits to this approach:

- This introduces better "navigational consistency", meaning that every
  launch of PageActivity will now be a proper self-contained browsing
  "session".
- The user will still be able to navigate to links and build a backstack
  in the current tab, but the number of Back presses to return to the
  parent Activity will now make intuitive sense, since the newly-opened
  page will not be pushed onto the existing backstack of the current tab.
- Notice that when the app is launched from an external Intent (e.g. from
  a deeplink) we do already open the requested page in a new tab,
  precisely for the purpose of maintaining consistent Back behavior. We
  should simply treat "internal" Intents coming from MainActivity the same
  way we treat external intents.

This patch also limits the total number of possible tabs to 100, since
we're technically leaving open the possibility for the user to create more
tabs than they normally would, if the user always uses the home-as-up
button for backwards navigation instead of the hardware Back button.
Limiting the total number of tabs doesn't seem too unexpected, since users
will still be able to refer to their History and Reading Lists for reading
previous pages.

In addition, this patch removes a bit of dead code.

Change-Id: I861b2822942f28c95ac0c3d3fb6f56ed18a8fd00