Retrieve group membership info when logging in
This is in preparation for bug T132834 (Semi-protected page cannot be edited).
The problem is that with the highly cached responses we get from RESTBase
the app doesn't get any user specific responses anymore.
In the case of a semi-protected page, it would return false, no matter as
what user one is logged in. Instead of taking that as the truth we could
consider the value of "editable" in the lead response to be "editable for
anonymous users". To be able to do that we need to retrieve the group
membership information for a logged in user first.

The membership info would only be available for newly logged in users.
If it's important that it works for already logged in users we could follow
up with another patch which checks if there is membership data when
downloading a page.

GroupMembershipClient:
Added another DataClient which automatically gets executed
when the login process is complete.

The login would be marked as failed if the retrieval of group membership
info fails. We could theoretically mark it as successful instead. This is
just so we can say that if a user is newly logged in and has the proper
membership she should be able to edit semi-protected pages.

The list=users request could provide more information in the future.
We just don't need more at this time.

Bug: T132834
Change-Id: Iccb8a482bbe4646ed3f625b7da8e3284a5a3f8d8