Check user group memberships on protected pages
Now that we know what groups a user is a member of we can check
if the user is a member of any of the allowed groups to edit the page.

This will enable a logged in user to see the regular edit pencils instead
of the disabled ones. The server still ultimately checks permissions upon
an edit submission.

Bug: T132834
Change-Id: I01a180f256f3a82ac197ed23445e976581ca8813