Simplify loading of styles.
This simplifies the way that we load CSS styles into our WebView
by simply including them in a <link> tag within the <head> element in
the HTML file, instead of loading the CSS via a Bridge event. This was a
vestige of the time when the app used to dynamically download CSS
bundles in the field, and is no longer necessary.

Change-Id: I481d5102ccffca407178e6b98b3e5d2c8363809b