Always show "search" and "tabs" buttons in Explore and Page toolbars.
Possibly controversial, but hear me out:
Our current menu for the Feed has all items marked as "ifRoom", which is
strictly correct by Android guidelines. However, on small-ish screens, it
can cause some of the menu items to overflow into the native overflow
menu, which we don't actually want. I can reproduce this easily on an
emulator with a small enough screen. (The overflow sometimes happens
even when there's seemingly enough space for all the buttons.)

It's very important for the user to be able to access the "search" button
in the toolbar at all times, as well as the "tabs" button. If these
buttons overflow, it greatly reduces their discoverability and
accessibility. This issue has already been raised via OTRS feedback.

Therefore, in this rare case, I propose we mark all of these menu items
with showAsAction="always", and add a rule to ignore the corresponding
lint warning.

I'm also marking our custom overflow button as "always", because if it
overflows, it will just create a native overflow button, which defeats the
purpose of our button, and takes up an equivalent amount of space.

Change-Id: If102252a9ce30e8a62f5650ad63037815624c712