Fix possible crash on Samsung devices.
On Samsung devices with API 23+, the floating ActionMode can cause a crash
when exiting an activity while the ActionMode is active (the system seems
to be attempting to attach the same ActionMode to the new activity).
(This is now our most prevalent crash in production, not counting the one
that was fixed earlier.)

This patch explicitly closes any ActionMode when pausing PageActivity.

As a bonus, this also seems to fix a potential leak of a Window object on
other (non-Samsung) devices when exiting the activity with an active
floating ActionMode.

Bug: T147191
Change-Id: I8cf8f803aec81cfe77612f81b683d4e55c5442ef