Correctly handle Site language when instantiating Retrofit clients.
The switch to using a single instance of OkHttpClient left out a crucial
bit of information: the Site object contains a languageCode field that is
used to determine custom headers (for language variants) that are sent
with each request, whereas the new single-instance OkHttpClient was simply
using the subdomain name of the Url as the language code.

This restores the Site object as a parameter for making new Retrofit
instances, and preserves the Site for use by the Interceptor that adds
custom headers.

Change-Id: Ic40d795746d9722c490407813228273fb2c13ced