Make sure daily recurring tasks are executed when MainActivity is run.
Right now our recurring tasks are triggered only when PageActivity is
started, which was acceptable in the previous incarnation of the app.
However, we now have MainActivity, with which the user can interact
without ever launching PageActivity, meaning that the recurring tasks
might not get executed.

This can actually mess with our engagement/retention metrics, since one of
the recurring tasks is DailyEventTask, which sends an event that we use to
measure retention.

This patch moves the execution of recurring tasks to the base activity
class, so that the tasks are executed when *any* of our activities is
started, not just PageActivity.

Change-Id: Ic9d41817740a65aac4ade0840efe784b32603b87