Fix possible crash due to null CookieManager.
Although I've been unable to reproduce it, I'm fairly sure about the cause
of this crash:

Inside OkHttpConnectionFactory, we initialize the CLIENT statically.
However, the instance of the client depends on the cookieManager to be
initialized in WikipediaApp, *but* the initialization of the cookieManager
only happens in onCreate(). This leaves open a possibility of CLIENT to be
initialized while cookieManager is still null.

This patch switches to initializing the CLIENT lazily, ensuring that the
cookieManager will have been initialized prior to any access of the
client.

Bug: T148172
Change-Id: Ieb16b0f4b9ea85eeabfb6923417f9bc3d9a85b94