Add field checking for API responses
Provides a mechanism for enforcing the API contract on the client side.

Specifically, this provides a @Required annotation that can be used to mark
fields that should be considered required during deserialization.  When a
field in a class is marked @Required, a custom Gson type adapter will be
created during unmarshalling that will check to ensure that the field is
non-null before instantiating a new instance of the class, and otherwise
return NULL rather than a malformed object.

Tests to follow.  Implementation in the Gson POJOs to follow.

Bug: T145075
Change-Id: Ic4e9d62cb24cb453d1c34e0618ff83479a924fe9