Hygiene: update SaneAsyncTaskTest to JUnit 4
I thought this test was causing problems when I looked at the log but it
was actually expected to throw (and log) Exceptions. Upgrade to JUnit 4
and clean up code where easy. Also, use UI thread when referring to
caller since AsyncTasks require execute() to be called on the main
thread.

Change-Id: Ie7fb15a9837f8ff71e76531c9c19ec75f9841496