Upgrade to screenshot-tests-for-android v0.4.2
Upgrade to screenshot-tests-for-android v0.4.2[0] which advertises
"Support for Android gradle plugin 2.2.0". Also update all screenshots
for the new library and update CI emulator configuration. The wiki has
also been revised[1].

One screenshot was removed because the test was updated to remove a
configuration previously but the screenshot wasn't accounted for then.

[0] https://github.com/facebook/screenshot-tests-for-android/blob/master/CHANGELOG
[1] https://www.mediawiki.org/wiki/Wikimedia_Apps/Team/Wikipedia_Android_app_hacking#Screenshot_tests

Change-Id: I2b3d7855cf2ec130ba44cceaf3a18ffd88086eb6