Create embryo description editing activity.
This makes article descriptions clickable (by adding a ClickableSpan to
the relevant text portion).
This also creates a blank activity that will be the container for the
description editing workflow. The activity is not entirely blank: it
accepts a PageTitle parameter, which feeds it with the requested page
title, which will contain its current description.

Bug: T146639
Bug: T146640
Change-Id: Iaa62a901d93376cf743bfd933db0b6ea47be658c