Fix bug in Namespace to legacy string conversion
Fix capitalization bug in Namespace to legacy string conversion and add
test. When converting a non-main Namespace to a legacy string, the
expected result is sentence case with underscores. For example, "Talk"
or "Gadget_definition". Fix a state bug causing the result to be fully
capitalized like "TALK" or "GADGET_DEFINITION".

Change-Id: I383f5728cfb551181d0ba9e41657863ce7cda9f4