Fix picture of the day caption being single-line.
I've noticed that the caption shown on top of the picture of the day is
overflowing onto multiple lines, despite having the attribute of
maxLines="1". This seems to be fixable by removing the textIsSelectable
attribute, which isn't particularly necessary, since the user may click
the card itself, and select the text from the full-screen gallery caption.

Change-Id: Ife44dd4f48e8be93d0605f8fc137d899f9d739f0