Remove ambiguity when using GoneIfEmptyTextView.
I've noticed that in the current Feed, we have a regression:
In the list of trending articles, if an article does not have a
description, the title should become vertically centered, whereas it now
incorrectly stays near the top of the list item.

This is because the description field is a GoneIfEmptyTextView. This
component very deceptively exposes a setText(String) function, which
does NOT override the super function setText(CharSequence). This can lead
to all kinds of subtle visual bugs like this.

This patch finally patches this inconsistency once and for all, by making
this component use a TextWatcher, and update its visibility based on the
text content that is fed into it.

Change-Id: I0d1675b801b17df36aa5d61d9200d596e1c35c1d