Upgrade to Support Library v25.0.0
Upgrade the Support Libraries to v25.0.0[0], buildToolsVersion to
v25.0.0[1], and compileSdkVersion to 25. There were several bugs
mentioned in the Android subreddit[2,3,4,5] but none I could repro in
the app. There was one noteworthy new feature added mentioned in the
release notes that is not tracked in Phabricator:

  The new DiffUtil class can calculate the difference between two
  collections, and can dispatch a list of update operations that are
  suitable to be consumed by a RecyclerView.Adapter.

The paradigm for Gradle dependencies is unclear. The Topeka[6] example
doesn't match the library breakdown by feature[7].

The upgrade does not seem to have affected any of the Spoon or unit
screenshot tests.

[0] https://developer.android.com/topic/libraries/support-library/revisions.html
[1] https://developer.android.com/studio/releases/build-tools.html
[2] https://www.reddit.com/r/androiddev/comments/500cx9/android_design_support_2420_update_password/
[3] https://www.reddit.com/r/androiddev/comments/52m6ac/android_support_library_revision_2421_released/
[4] https://www.reddit.com/r/androiddev/comments/4y70e7/android_support_library_v242_released/
[5] https://www.reddit.com/r/androiddev/comments/58czcd/support_lib_25_with_bottomnavigationview/
[6] https://github.com/googlesamples/android-topeka/blob/ed5573691843d6afde4086f86ed2f7659016ec9b/app/build.gradle
[7] https://developer.android.com/topic/libraries/support-library/features.html

Bug: T144318
Change-Id: Ibdf76a6b640b5cb749caaf2b3e4bb3186d6fa710