Fix possible leak in FacePostprocessor.
Observed in API 17, this leak seems to happen because the
FacePostprocessor class implicitly retains an instance of the
FaceAndColorDetectImageView of which it is a member.

This patch separates out the FacePostprocessor class, along with its
concerns, and makes sure that it doesn't keep any references to the View
that uses it.

Change-Id: I6e637b4e5fe676f9b6bc13f1e2ff21d1c500bbaa