Fix possible memory leak(s) in feed.
Observed in API 17, certain Clients that provide feed content seem
to retain a reference to a Context, which in this case is MainActivity.

This patch provides the FeedCoordinator with an Application context
instead of an Activity context, so that it won't matter if the coordinator
(or any Clients to which it passes the Context) holds on to the Context
longer than necessary.

Change-Id: Ia2603d0ccf57889d84a81d1b90f71cc88f78884d