Fix possible crash when clicking on an image.
Based on a (relatively small) number of crash reports, it looks like some
articles have images that link to their subjects without the "/wiki"
prefix, which doesn't match the expectation of our imageClicked event
handler. This causes the handler to pass the link off to the parent
LinkHandler, which causes it to look for a "title" attribute in the
payload, which doesn't exist, which causes an exception.

This patch makes the "title" attribute in this case optional, and also
hardens the handling of the "title" parameter within the LinkHandler
itself, so that if the titleString happens to be null, the original href
is used.

Bug: T149692
Change-Id: I0b2a35db20cbb35c5b557b4fdeb9966ae6a3f030