Fix visual nuisance when adding to a reading list from the feed.
Take the following steps to see what I mean:
- Open the feed, and scroll down so that the navigation bar becomes
  hidden.
- Tap the button to save one of the displayed items to a reading list
  (e.g. tap the "save" button on the Featured Article).
- Select a reading list to save into, and observe a snackbar that gives
  you the option to "View list".
- When you tap the snackbar, it takes you to the reading lists tab, and
  the topmost reading list is weirdly scrolled underneath the main
  toolbar.

This patch adds some code that programmatically scrolls the
CoortinatorLayout just enough so that the bottom navigation bar is fully
shown, and the inner content of the current tab is fully scrolled into
view.  Ideally we probably wouldn't want the reading lists screen to be
"scrollable" when there's no scroll-worthy content in it, but that's for
later.

Change-Id: I9f82f6d166d778bb57fe7a54a873b82bc4ca8b96