Fix long-press menu actions in search results in MainActivity.
I've found that when you invoke the Search interface from MainActivity
(i.e. the feed), some of the long-press actions don't seem to work,
including add-to-reading-list, share, and copy link.

This is because the function to obtain the callback() object for the
current fragment only looks at the "parent" fragment or "target" fragment,
whereas the Search fragment is actually a *grandparent* of MainFragment.

This patch refactors the Callback interfaces of the search sub-fragments
so that they're only directional with respect to the immediate parent of
the fragment.

Change-Id: If71346f4c77cca0e7793ad61fc05ceb15babe2bf