Fix a few annoyances in Nearby.
This patch nails down some logic that hasn't yet been optimal when
navigating and switching back and forth to the Nearby tab.

- Currently, when you open the app and go to the *History* tab, the app
  sometimes shows a snackbar that actually belongs to the Nearby tab which
  says "location services are disabled", which is not true.
- Currently, when pressing the "my location" button for the first time
  (and accepting the location permission), the app still issues a snackbar
  that says "location services are disabled", which is incorrect.
- Currently, when switching from Nearby to Explore and back to Nearby, the
  previous location and results are not preserved.

This patch fixes these inconsistencies by introducing a more proper
function for determining whether location services are actually enabled,
and also improving the logic of persisting the current location of the
mapView, as well as the logic of when to jump to the user's location.

Change-Id: Ia43f5254b755b2a424e38af54e78c964e1339880