Echo notifications for description edits.
This introduces logic to show a system notification when an edit made by
the user is reverted by another user. This uses the Echo API to retrieve a
list of currently-unread notifications, and then issue a request to mark
them as read.

Technically, this logic is capable of showing arbitrary Echo notifications
for any sort of reason, but is currently limited to Wikidata deliberately,
until we decide to enable notifications for other reasons.

Also, something specific to Wikidata:
For notifications that come from Wikidata, the "target" article in the
notification is represented as the Q-number, instead of the human-readable
article name. Therefore we need to make an additional request to resolve
the Q-number into the correct label for the user's language. I haven't yet
found a way to get notifications from Wikidata with the Q-number
pre-resolved.

Important: This requires a new permission to receive the BOOT_COMPLETED
broadcast, so that our notification service can start upon boot, without
needing to launch the app. (Previously the BOOT_COMPLETED permission was
only for preload partner builds.)

TODO: Limit NotificationPollService to run for only one day after the user
makes an edit (or alternately run all the time, but only make network
requests for a day after an edit is made).

Bug: T150477
Change-Id: Ic2ab7bc402064cb8bf7351f6482b5b27a8925b03