Fix/restore Wikipedia Zero enter notification.
For some reason, the notification when entering a Zero network seems to
have disappeared (?), along with the code that produces it. This seems to
have been due to https://gerrit.wikimedia.org/r/317166

This brings back that notification, and also adds separate icons for
entering and leaving a Zero network.

Bug: T150258
Change-Id: I21d556b0ea4f3fae27f4d98e4d11f0ee9674763f