Fix server logout case when logged in user edits
When a user is logged into the app and makes an edit, EditClient
requests that the server verify the user is still logged in and that the
edit will be associated with their account. If the user is logged out,
the request fails. This patch checks for and retries in that scenario.
I no longer reproduce this issue but I wrote a test to exercise the
scenario.

Bug: T150529
Change-Id: I88b700b2bfb9433ddcaab86829b32557000f772d