Store time spent reading in History entries.
This improves the smartness of our History database by adding a "time
spent" column, which keeps track of the total time the user spends looking
at the article. This field is cumulative: if the user visits the article
multiple times, the reading durations are added together.

This also makes use of this new column by plugging it into our Feed
content: the Continue Reading and Because You Read cards will now only
show suggestions for articles that have been read for 30 or more seconds.

Bug: T146646
Bug: T146647
Change-Id: Id8d198a3340c58f7077bd46ebc6147545143d4c4