Fix MediaDownloadReceiver to work with all API versions.
This resolves one of the blockers for targeting API 24+.

This fixes the way we update the device's MediaStore upon completion of a
downloaded image or other media file. Instead of using the deprecated
COLUMN_LOCAL_FILENAME, we now use COLUMN_LOCAL_URI. All of the other logic
remains unchanged.

Also, it looks like the MediaDownloadReceiver wasn't wired correctly into
MainFragment, so it wasn't receiving the download completion event. This
patch wires it in correctly.

Bug: T144920
Change-Id: Ie760cd7f86f52a18d0c63ad658f73eba8f0e7c1e