Persist channel parameter from install referrer.
This introduces a new parameter for our install referrer string, called
"channel". If this parameter is present, we will save the channel in
SharedPreferences, so that it will be passed through to the userAgent that
the app uses for all network requests. (The logic of putting the channel
into the userAgent is already there.)

Bug: T149917
Change-Id: Ia4ea46765dacde8e1c1957dc6f94bd36d5807462