Refresh login session for Wikidata editing
When the login token/cookies expire we get an error from the server
when executing the first request.

When this happens we refresh the login tokens then retry
the first request one more time.
Unfortunately, I did not save all the logs from when it happened earlier.

Theoretically there is a small window of time where this
could happen on later requests as well. Not sure if it's worth adding
more complexity to this at this time.

There is room for refactoring to share common behavior with the
page edit scenario. See I88b700b2bfb9433ddcaab86829b32557000f772d

Change-Id: I3d8d7e7220f80d1cb0a7181afef1edb38518a7ef