Capitalize Wikidata description in the presentation layer only
Fixes issues caused by the Wikidata description being transformed to
capitalized form when read, including being presented for editing in
capitalized form and unintentionally modified to be capitalized when
saved.

Unfortunately, doing this in xml with android:inputType="textCapSentences"
doesn't work, and TextView.setText() is final and can't be overridden to
do this in a custom view, so calling StringUtils.capitalize when setting
the description text on a view looks like our best option.

Bug: T151336
Bug: T151337
Change-Id: I7c6e8de21ab41fc9296f2f6582d4dbabd9b65cfd