Hygiene: Separate Activity classes into themed and unthemed variants
The DescriptionEditSuccessActivity is an instance of a situation in which
we don't want some of the theming assumptions currently made in our
Activity class hierarchy.  For instance, SingleFragmentActivity sets both
a background color (in activity_single_fragment.xml) and a theme (as a
result of subclassing ThemedActionBarActivity) that were unwanted for
DescriptionEditSuccessActivity.

This patch removes the theming logic from the main branch of the Activity
classes and creates Themed variants of both the ActionBarActivity and
SingleFragmentActivity for when the default theming logic is desired.

No functional changes intended.

Change-Id: I3c2d07aa04c0f04a9f1d07a1c2fc72b410ab8ded