Copy CentralAuth tokens for www.wikidata.org requests
Copy CentralAuth tokens for www.wikidata.org requests from wikipedia.org
if the user is logged in. There are tokens that start with "centralauth_".

Examples:
* centralauth_Token
* centralauth_User
* centralauth_Session

Gergo mentioned that this would create the local session on the fly.
This would allow us to skip CentralAuthClient implementation we had added earlier
and reduce the number of requests per description edit.

This happens independent of the data client. It's all contained in the
SharedPreferenceCookieManager.

The one thing added to the data client is to now add the assert=user if
a user is logged in.

Bug: T150705
Change-Id: Ic3e05370d9bb028bac80446acf8b184c49b479a7
(cherry picked from commit c5400639bb916e44b50b32e66113464db2bfa821)