Generalize Announcement cards a bit.
The previous (first) version of announcement cards didn't actually make
use of the "caption_HTML" or "image" parameters that are part of the
announcement endpoint of the Content Service, which are meant to represent
the footer verbiage and the header image, respectively.

This patch updates our logic to make use of these fields, which means that
we no longer need to maintain any hard-coded images or strings for this
purpose.

We will still, however, keep around the separate classes of SurveyCard and
FundraisingCard, since this distinction is necessary for analytics
purposes. But aside from that, the card is now fully customizable
by the endpoint content.

(Also update the Test to be a more general test of AnnouncementCardView.)

Change-Id: I8e77fc7bcbdd9ed020127f40c1996ff772fff097