Fix possible crash when visiting external link.
Our method to visitInExternalBrowser() was actually making an assumption
that the Context being passed into it is really an Activity, when in fact
it could be an Application context. The crash occurs because the system
doesn't want to create an Intent without a backing Activity when the
Intent doesn't have a flag of FLAG_ACTIVITY_NEW_TASK.

This patch adds that flag, which fixes the crash when the Context is not
an Activity, and has no effect when the Context is an Activity.

https://rink.hockeyapp.net/manage/apps/226649/app_versions/113/crash_reasons/148648041

Change-Id: I18c66bda32315d4ef97cf15a61cf7727fa7ec805