Fix possible crash when resuming activity.
On certain Samsung devices, in certain cases, the invocation of onPause()
doesn't seem to happen symmetrically with the invocation of onResume(),
leading to an imbalance of our calls to register/unregister with the Bus.

This patch moves the register/unregister calls to onCreate()/onDestroy(),
which seems to be more consistent on these devices.
(repro'd, with difficulty, on my Galaxy S7.)

Change-Id: I4f923c1008da74e74be767a331c4130ae581f768