Fix possible crash when edit preview times out.
https://rink.hockeyapp.net/manage/apps/226649/app_versions/113/crash_reasons/148711407

The "failure" callback of the edit preview request was checking whether
the caught exception was an instance of ApiException, which was OK prior
to switching to Retrofit, since ApiException used to encompass socket
timeouts that happened within the java-mwapi library. But now, the caught
exception can itself be a SocketTimeoutException, among others. The check
for ApiException is no longer necessary.

Change-Id: I2a79a9ca9d6b7454c13e6971f826b2d3bc2f7301