Hygiene: Clean up misuse of PageTitles in GalleryActivity
The logger expects a PageTitle object coming from the gallery activity,
but we may not be in the context of wiki page since the gallery can now
also be launched from the feed featured image card.

An earlier commit introduced the idea of keeping around a placeholder
PageTitle for logging purposes in the event we're coming from the feed
featured image card.  But storing this in the pageTitle member variable
is a bad idea since this makes creates a risk it'll end up getting
inappropriately passed to the content model for the page data client or
inserted into the cache.

This patch eliminates such use of the PageTitle class and updates
downstream usages of the PageTitle object as need.

To follow up: figure out a good way to pass the featured image date
through to the gallery for use in a share subject.  A
FeaturedImageGalleryItem subclass would likely be appropriate here.

Bug: T152980
Change-Id: I74ef3557700018176310097180515664652b13ca