Fix size-limiting issue in PageCache.
This fixes a bug that was preventing our PageCache from being trimmed to
our specified maximum size of 64MB, and instead was allowing it to grow
indefinitely.

This is fixed by calling flush() on the cache *after* calling commit() on
the editor object, instead of before.

Upon examining the source code of DiskLruCache, we can see that it enters
a loop when trimming the size of the cache, and within the loop it evicts
old entries until the size is less than the maximum.

HOWEVER: If the editor object wasn't committed yet, the evict() operation
will be unsuccessful, and the size of the cache will be unchanged, which
will send the trim operation into an *infinite loop*, making this an
eyebrow-raisingly serious issue.

Change-Id: I6bdfca20e8b751b3a94a0b0c234ba31c715a0e3b