Fix possible crash when upgrading from older app versions.
In versions of the app prior to 2.1.142, the list of Tabs was serialized
in a way that's not completely consistent with the way it's serialized
today. Specifically, the "domain" item of the "site" component can
sometimes be an empty string, which would manifest itself as a Uri with
an empty authority.

This leads to a dangerous condition where the tab list is deserialized
"successfully", but the Uri contains a ticking time bomb of null members,
which causes a confusing exception when the Uri is passed to Retrofit.

This patch adds logic to look for tab entries that match this description
after deserializing, and if found, fall back to clearing out the tab list
and starting fresh.

Bug: T152980
Change-Id: If09f6832e6a4f3e2b47a3806cb4e8da9cad8031e