Fix scrollability of article when tapping on description.
Since we endow the description with a LinkMovementMethod, it no longer
passes scroll events down to the WebView, making it impossible to scroll
the article by dragging the description. This is more of a problem now
that we're allowing descriptions to span multiple lines.

This patch sets a LinkMovementMethod only if the description is empty
(only if we actually want the description to be clickable), and otherwise
sets it to null.

Change-Id: Ie462560bf82e9ac43b5d0891d2b17ef8304240ac