Fix crash caused by vector drawable on pre-lollipop devices
Switch to create a VectorDrawableCompat of the Nearby map marker icon to
give to the Mapbox IconFactory, rather than passing a raw resource id,
which crashes on pre-Lollipop devices.

This crash occurred when the user selected either Nearby or History from
the bottom tabbed navigation menu.  Apparently the ViewPager was getting
resources for the adjacent views to the current view teed up in the latter
case.

Bug: T153968
Change-Id: Iecbe0be5756d9bc9e34b119fdeb81dabb2fe297e