Description editing tutorial: further design tweaks
Moves the PageIndicatorView from the fragment layout to the view layout.
This has two benefits:

(1) we no longer require an offset that can potentially obscure text at
    larger text size configurations;

(2) the page indicator view is included in the screenshots used for
    testing.

Further tweaks the view layout to accommodate this new arrangement.

Sadly, this means that the page view indicator will no longer remain
stationary while scrolling between the pages but will appear separately
on each page; such is life.

Bug: T148205
Change-Id: Ic443210ef8a6512d19372eaef7a9c99e0a758f46