Make WikitextClient expect a MwQueryResponse<Wikitext>
WikitextClient currently can't adequately detect and handle MediaWiki
API errors because its requests are for bare Wikitext objects with no
notion of an error response, rather than a MwQueryResponse<Wikitext>.
This changes it to use the latter.

This should fix any crashes, but the underlying cause in T152982 is
still unknown.  An error can occur if the section the user requests
disappears between loading the page and clicking edit, but that doesn't
actually appear to be what's going on here.

Also, our error handling in the section editing activity is inadequate
and will need updating for whatever resolution we decide on. (T154805)

Bug: T152982
Bug: T154807
Change-Id: I1955924c8b3230a536acaa9b44218db7ca3a663b