New: make it harder to be a single lingual dev
This patch may be controversial because it makes it harder to program
blissfully unaware of other languages. The intent is to force devs to
experience the app in other, especially non-English, locales regularly.
The mechanism used is to pick a random language on initial app install
of the dev flavor.

Change-Id: Id86349d230b521ddc56e9f1c21d0a57b25bc210a