Hygiene: Consolidate MwApiResultPage into MwApiResponsePage
MwApiResponsePage is a model class meant to model a standard page object
as returned by the MediaWiki API.

MwApiResultPage is also intended as a model class for the same purpose,
but also includes some incidental feed-related cruft.

This consolidates MwApiResultPage into MwApiResponsePage and moves the
cruft to various more-appropriate places.

Change-Id: I039737f326626724ff4affc979b2c72b1073224c