Make use of IME action button when editing descriptions.
This enables a "Done" button on the soft keyboard when editing a
description, which takes the place of the carriage-return key. Clicking
the button will have the same effect as the Save button in the toolbar.

Unfortunately it doesn't seem to be possible to dynamically "enable" and
"disable" the Done button inside the keyboard in the same way that we do
with the Save button, so the keyboard button is always enabled, but it
will simply have no effect if the Save button is disabled (i.e. if the
description is unmodified or empty).

This also enables long-press hints for the buttons in the description
editing UX.

Bug: T150712
Change-Id: I00f8b185eb671e75de7d9aaee4c9c37b5de1aec7