Get user ID after logging in or creating an account
The MW API doesn't return user id on login or account creation success.
We have to get it separately.  Currently, we don't do this, and the user id
in the app is always 0.

It seems the numeric user ID is needed for EventLogging, so we'd better
get it.

For the login case, this adds another paramter to the request that's
already made after login to retrieve group memberships, so that the user
ID is also returned.

For account creation, a simple new UserIdClient is added to make the
follow-up request.

Bug: T149915
Change-Id: I3dc7871805edf51d2de6598dd5a5e67140f416b5