Show CC license disclaimer when editing descriptions.
Includes a couple of design tweaks in DescriptionEditView.

This also updates the tint of the CC icon in the section editing activity,
and updates it to use the vector drawable.

TODO: Factor out the display of the CC logo + license disclaimer into its
own View, to be shared between the Section editing and Description editing
activities.

TODO: Remove the raster CC logo image, which is currently only being used
by some screenshot tests.

Bug: T156279
Bug: T153483
Change-Id: Id501b2fb337abe4ae394d68753d36368f60a3188