Show notifications in heads-up style.
This configures our notifications to appear as "heads-up" notifications,
meaning that they will pop up on top of the current app, in addition to
just showing a notification icon in the system bar.

This applies to Revert notifications, as well as Zero enter/exit
notifications.

The key was to set the DEFAULT_ALL flag on the notification, along with
the PRIORITY flag being set to high or max.

Change-Id: I24c8f54a5c6c6761e7ddcd5ad87469d751f0fcac