[Facepalm] Correctly set interval for notification polling.
In my ongoing testing of our notifications, I've been noticing that our
polling task doesn't actually repeat like we tell it to. After hours of
fruitless debugging, I noticed that we were passing in the raw integer
resource ID, instead of the actual integer dimension. Therefore, we were
inadvertently setting a repeat interval of 2132897234 or something
similar.  The same for the number-of-days timeout for the polling task to
stop itself.

This patch switches to the much-more-correct scheme of using the
getInteger() method of retrieving the integer dimension, instead of using
the resource ID.

I suspect that our "testing" of notifications has worked so far because
after editing a description we call startPollTask() which effectively
queues the task for immediate execution.

(This patch also cleans up the code a bit.)

Change-Id: I5e0f6b0bb3813763133f891b9475abb232dee62a