Remove IME action when editing a section.
When editing a section (on API 25), the soft keyboard was showing a "Done"
button, which not only doesn't do anything, but it also prevents the user
from inserting a carriage-return into the text.

This explicitly sets the IME action to None. (It doesn't seem to be enough
to simply leave out the "imeOptions" xml attribute...)

Change-Id: I9ecd3e855fa06ac2fbae20487de1dd39954ea01d