Convert link preview dialog to be a native BottomSheet.
This makes our link preview dialog become a true BottomSheet, which the
rest of our bottom sheets already are.

The only thing we "lose" by doing this is the overlay "Read article"
button that floats on top of the sheet (the button will now be fixed at
the bottom of the sheet). However, the user may always tap the title
header of the link preview to navigate to the article.

The benefits of this change are:
- This was one of the last missing pieces for multi-window support on
  Samsung and Android Nougat devices.
- A great simplification of the code, and removal of custom reinvention
  of bottom sheet functionality.

TODO: Make the link preview dialog use butterknife.

Bug: T109341
Change-Id: I6290e14774a5746c6a0988276139cfcb8cbd2978