Fix/improve page backstack behavior.
Our tab/backstack logic is some of our more convoluted code to wrap one's
head around (sorry), but here's what this patch does:

- Previously, when PageFragment became attached to the activity, it
  automatically (unconditionally) asked PageDataClient to load the last
  page that was on the backstack. This is actually not quite correct,
  because the very next thing that the activity does is tell the fragment
  to load a new page. So what happens is that PageFragment pushes one page
  onto the backstack implicitly (without the activity telling it to), and
  then pushing another page onto the backstack explicitly when the
  activity tells it to load the next page.

The above logic may have actually been degrading the app's performance by
thrashing the WebView, by starting to load the implicit page, and
immediately following by loading the new page. This effect wasn't
perceptible during testing, which is why it was overlooked.

This patch adds some conditions to PageFragment's implicit loading of the
last page on the backstack. Namely, the only reasons for implicitly
loading the page is (a) when restoring from savedInstanceState, and (b)
when PageFragment is invoked to show the list of tabs.

This also simplifies some of the logic around loading the Main Page, and
when/how this actually happens (some of the logic is no longer necessary
due to our overhauled navigation).

And this also moves the logic of checking whether the requested page is
already being shown to PageFragment instead of PageActivity.

Note that this patch shouldn't be changing any user-visible behavior with
respect to tabs or navigation. On its face, it simply fixes the
above-mentioned backstack duplication bug that was observed recently.

Change-Id: I492acbbb96daad03c12a0cdf4a6b8d30ad9051e2